Add ConversationEventMap, Callbacks, and event-name mapping to events.ts
Move Callbacks and CALLBACK_KEYS from types.ts to a new events.ts module
that combines generated events with hand-written internal and extra events
into a unified ConversationEventMap. Add CALLBACK_KEY_TO_EVENT_NAME mapping
for wiring callbacks to event names.
